NOJIRA Adjusted Humidity to be a Double instead of an Integer.
git-svn-id: https://source.jasig.org/sandbox/WeatherPortlet/trunk@15213 f5dbab47-78f9-eb45-b975-e544023573eb

+++ src/main/java/org/jasig/portlet/weather/dao/accuweather/dom4j/WeatherUtil.java
@@ -154,11 +154,11 @@ public class WeatherUtil {
 		}
 	}
 
-	public Integer getHumidity() {
+	public Double getHumidity() {
 		if (currentConditions != null) {
 			Element currHumidity = currentConditions
 					.element(Constants.CURR_HUMIDITY_TAG);
-			return (currHumidity != null) ? Integer.valueOf(currHumidity
+			return (currHumidity != null) ? Double.valueOf(currHumidity
 					.getText()
 					.substring(0, currHumidity.getText().indexOf('%'))) : null;
 		} else {
